requirements.txt
requests
pymongo
faktory
pyfaktory
python-dotenv
matplotlib
pandas
seaborn
colorlog
numpy
